Fix integer overflow.
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/trunk@1396088 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/codecs/src/java/org/apache/lucene/codecs/compressing/CompressingStoredFieldsIndex.java
@@ -363,7 +363,7 @@ public enum CompressingStoredFieldsIndex {
         throw new CorruptIndexException("Expected " + numChunks + " chunks, but got " + startPointerDeltas.size());
       }
       final PackedInts.Mutable startPointers = PackedInts.getMutable(numChunks, bitsPerStartPointer, PackedInts.COMPACT);
-      int startPointer = 0;
+      long startPointer = 0;
       for (int i = 0; i < numChunks; ++i) {
         startPointer += startPointerDeltas.next();
         startPointers.set(i, startPointer);
